Do not nullify the extra data from the config (#36)
For some reason the installer was assuming that if you dont do an
automatic installation, your "extra data" from the config file should be
gone.

I do not understand why as automatic is not the same as providing a
config file with values which may have nothing to do with the
installation.

This patch removes the nulling of the extra data

Signed-off-by: Itxaka <igarcia@suse.com>