Use ros-operator image for the ros-operator chart
Create a simple ros-operator image with only the ros-operator binary on
it as it should be enough.

Use it as default for the chart.

Introduce REPO_ROS_OPERATOR to point to the proper repo to store the
ros-operator.

Set the REPO_ROS_OPERATOR to ros-operator-ci for ci and ros-operator for
release

Signed-off-by: Itxaka <igarcia@suse.com>