Drop rancheros-operator from os2 repo
It is now living on https://github.com/rancher-sandbox/rancheros-operator
Drops the chart as well.
Modifies all scripts so they go the proper way of installing a chart,
i.e. adding the repo and installing from the repo directly.
Modifies dockerfile to still get the chart into the image as before.

Signed-off-by: Itxaka <igarcia@suse.com>