Do not use cdrom provider datasource (#212)
* Do not use cdrom provider datasource

Avoid the use of 'cdrom' provider for datasource plugin. This provider
opens all available block devices to probe them, however an `lsof` call
show they are not closed afterwards. This might cause issues if within
the same process we pretend to repartition a disk for installation.

This is a workaround for rancher/elemental-operator#50

Signed-off-by: David Cassany <dcassany@suse.com>

* Completely remove datasource plugin

We are not targeting public cloud infraestructure for now, hence there
is no need to include such a stage.

Signed-off-by: David Cassany <dcassany@suse.com>