Cleanup parts referring to outdated stack
This commit mostly removes references to outdated/dropped projects such
as rancherd, ros-installer and ros-operator. In addition it sliglty
adapts the surrounding text not to be completely wrong.

Note this is not a rewrite. It just tries to fix obsolete and even
wrong text, it is not aiming to provide missing contents.

Signed-off-by: David Cassany <dcassany@suse.com>