Cleanup parts referring to outdated stack (#249)
* Cleanup parts referring to outdated stack

This commit mostly removes references to outdated/dropped projects such
as rancherd, ros-installer and ros-operator. In addition it sliglty
adapts the surrounding text not to be completely wrong.

Note this is not a rewrite. It just tries to fix obsolete and even
wrong text, it is not aiming to provide missing contents.

Signed-off-by: David Cassany <dcassany@suse.com>

* Update docs/customizing.md

Co-authored-by: Francesco Giudici <francesco.giudici@suse.com>
Signed-off-by: David Cassany Viladomat <dcassany@suse.de>

* Adress some comments and keep Elemental Teal

Signed-off-by: David Cassany <dcassany@suse.com>

Co-authored-by: Francesco Giudici <francesco.giudici@suse.com>