ci: Add test for elemental-iso-build on PR
Now that we can modify where the artifacts are obtained from in the
script, we can tests under a PR context the validity of the script.

Also moves the script under the .github dir as it should not be on a
public facing place. Releases will be used to generate a script with the
proper tag embedded so it uses the proper tag to obtain the artifcats

Signed-off-by: Itxaka <igarcia@suse.com>