Add support for Podman for registration injection into ISO
Here I've added support for using Podman for the injection of registration
information to the Elemental Teal ISO. This does not replace Docker as the
default (if it's available on the client) but provides an option for those
of us using Podman instead (bypassing the `alias docker=podman` hack). This
has been tested on Tumbleweed with Podman and Ubuntu with Docker.

I have also cleaned up a little bit of whitespace that I found lingering.