Do not expand COS_PERSISTENT by default
This commit drops the expansion cloud-init step for COS_PERSISTENT
partition. The installation already uses all available size by default
there is not need to attempt a resize on every boot.

In addition this commit also updates the persistent volumes
configuration to be aligned with the latest elemental-toolkit
changes. Now volumes are set as part of the kernel arguments at boot
and not as part of the cloud-init setup. This prevents including
filesystem label literals within the immutable OS.

Signed-off-by: David Cassany <dcassany@suse.com>